Extract expressions into a module
Deleted the tests that were whitespace-specific when moving them
